adidas/requirements.txt
gradio==4.44.0
pandas==2.2.2
plotly==5.24.1
Pillow==10.4.0
openpyxl==3.1.5